Use kv store for checkpoint loading (#13282)
